Code refactoring. Take userID from JWT and update int type to int64

+++ internal/message/model.go
@@ -6,9 +6,9 @@ import (
 
 type Message struct {
 	ID         int64     `json:"id"`
-	SenderID   int       `json:"sender_id"`
-	RoomID     *int      `json:"room_id,omitempty"`
-	ReceiverID *int      `json:"receiver_id,omitempty"`
+	SenderID   int64     `json:"sender_id"`
+	RoomID     *int64    `json:"room_id,omitempty"`
+	ReceiverID *int64    `json:"receiver_id,omitempty"`
 	Content    string    `json:"content"`
 	CreatedAt  time.Time `json:"created_at"`
 	UpdatedAt  time.Time `json:"updated_at"`
